fix: some code organization
