+++ src/datalogger/src/datalogger/mod.rs
@@ -1,1 +1,2 @@
 pub mod settings;
+pub mod commands;
