+++ src/datalogger/src/protocol/mod.rs
@@ -1,2 +1,1 @@
-pub mod commands;
 pub mod responses;
